Added IM and EC3 services
Hi,

I've added IM and EC3 services to control the status of them. I have a doubt regarding EC3 portal, as you need to login with EGI Checkin before accessing the portal. If this is not possible, we can reference to the public endpoint of EC3 that does not need authentication (http://servproject.i3m.upv.es/ec3/).

Thank you!